requirements.txt
django<4.1.0,>=4.0.0
djangorestframework
pyyaml
requests
django-cors-headers
python-dotenv
mysqlclient
Pillow
